content-reco/requirements.txt
pyyaml
python-dotenv
google-cloud-bigquery
pandas
google-cloud-aiplatform
numpy
scikit-learn
flet
db-dtypes